+++ CHANGELOG.md
@@ -2,9 +2,16 @@
 All notable changes to this project will be documented in this file.
 
 #### 1.x Releases
-- `1.1.x` Releases  - [1.1.0](#110) | [1.1.1](#111) 
+- `1.1.x` Releases  - [1.1.0](#110) | [1.1.1](#111) | [1.1.2](#112) 
 
 ---
+## [1.1.2](https://github.com/congncif/IDMFoundation/releases/tag/1.1.2)
+Released on 2018-4-7
+
+#### Updated
+- String Key Value Protocol: Allow sending array in parameter
+- Moved parameterEncoding to BaseDataProvider
+
 ## [1.1.1](https://github.com/congncif/IDMFoundation/releases/tag/1.1.1)
 Released on 2018-4-6